+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Yii Framework 2 HTTP client extension Change Log
 
 - Bug #44: Fixed exception name collision at `Response` and `Transport` (cebe)
 - Bug #45: Fixed `XmlFormatter` unable to handle array with numeric keys (klimov-paul)
+- Bug #53: Fixed `XmlParser` unable to parse value wrapped with 'CDATA' (DrDeath72)
 - Enh #43: Events `EVENT_BEFORE_SEND` and `EVENT_AFTER_SEND` added to `Request` and `Client` (klimov-paul)
 - Enh #46: Added `Request::getFullUrl()` allowing getting the full actual request URL (klimov-paul)
 - Enh #47: Added `Message::addData()` allowing addition of the content data to already existing one (klimov-paul)